+++ couplet_composer/support/dependencies/lua.py
@@ -19,7 +19,7 @@ from ...build_directory import BuildDirectory
 
 from ...dependency import Dependency
 
-from ...invocation import Invocation
+from ...runner import Runner
 
 
 class LuaDependency(Dependency):
@@ -29,14 +29,14 @@ class LuaDependency(Dependency):
 
     def _download(
         self,
-        invocation: Invocation,
+        runner: Runner,
         build_dir: BuildDirectory
     ) -> str:
         """Downloads the asset or the source code of the
         dependency.
 
         Args:
-            invocation (Invocation): The current invocation.
+            runner (Runner): The current runner.
             build_dir (BuildDirectory): The build directory
                 object that is the main build directory of the
                 build script invocation.
@@ -57,23 +57,23 @@ class LuaDependency(Dependency):
         http.stream(
             url=download_url,
             destination=download_file,
-            dry_run=invocation.args.dry_run,
-            echo=invocation.args.verbose
+            dry_run=runner.invocation.args.dry_run,
+            echo=runner.invocation.args.verbose
         )
 
         source_dir = os.path.join(tmp_dir, self.key)
 
         shell.makedirs(
             path=source_dir,
-            dry_run=invocation.args.dry_run,
-            echo=invocation.args.verbose
+            dry_run=runner.invocation.args.dry_run,
+            echo=runner.invocation.args.verbose
         )
         shell.tar(
             path=download_file,
             action=ArchiveAction.extract,
             dest=source_dir,
-            dry_run=invocation.args.dry_run,
-            echo=invocation.args.verbose
+            dry_run=runner.invocation.args.dry_run,
+            echo=runner.invocation.args.verbose
         )
 
         return os.path.join(
@@ -84,7 +84,7 @@ class LuaDependency(Dependency):
     def _build(
         self,
         source_path: str,
-        invocation: Invocation,
+        runner: Runner,
         build_dir: BuildDirectory
     ) -> None:
         """Builds the dependency from the sources.
@@ -92,27 +92,27 @@ class LuaDependency(Dependency):
         Args:
             source_path (str): The path to the source directory
                 of the dependency.
-            invocation (Invocation): The current invocation.
+            runner (Runner): The current runner.
             build_dir (BuildDirectory): The build directory
                 object that is the main build directory of the
                 build script invocation.
         """
         with shell.pushd(
             source_path,
-            dry_run=invocation.args.dry_run,
-            echo=invocation.args.verbose
+            dry_run=runner.invocation.args.dry_run,
+            echo=runner.invocation.args.verbose
         ):
             shell.call(
-                [invocation.runner.toolchain.make],
-                dry_run=invocation.args.dry_run,
-                echo=invocation.args.verbose
+                [runner.toolchain.make],
+                dry_run=runner.invocation.args.dry_run,
+                echo=runner.invocation.args.verbose
             )
             shell.call(
                 [
-                    invocation.runner.toolchain.make,
+                    runner.toolchain.make,
                     "install",
                     "INSTALL_TOP={}".format(build_dir.dependencies)
                 ],
-                dry_run=invocation.args.dry_run,
-                echo=invocation.args.verbose
+                dry_run=runner.invocation.args.dry_run,
+                echo=runner.invocation.args.verbose
             )