+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for specifying both the source file and destination file when the installation of the dependency requires copying files.
 - Support for determining the asset file that should be downloaded for a dependency by using the key `asset`.
 - Support for determining the [GitHub](https://github.com) repository of a dependency by using the key `repo`.
+- Runner for each target and cross-compile target so that the script is run for every added target.
 
 ### Changed